requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0

# Google Sheets integration
gspread>=5.10.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
